Changed id types to longs and refactored to allow for easier integration testing, as well as implemented the first test for TransferDao

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -4,7 +4,8 @@ import java.math.BigDecimal;
 
 public interface AccountDao {
 
-    BigDecimal getUserBalance(int id);
+    BigDecimal getUserBalance(long id);
 
-    int getAccountIdByUserId(int id);
+    long getAccountIdByUserId(long id);
+
 }
